# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Speech Clinic API for children ages 2-5
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /specialists:
    get:
      operationId: listSpecialists
      tags: [specialists]
      summary: List all specialists
      responses:
        "200":
          description: List of specialists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialist"
    post:
      operationId: createSpecialist
      tags: [specialists]
      summary: Create specialist profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecialistBody"
      responses:
        "201":
          description: Specialist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
  /specialists/{id}:
    get:
      operationId: getSpecialist
      tags: [specialists]
      summary: Get specialist by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Specialist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
    patch:
      operationId: updateSpecialist
      tags: [specialists]
      summary: Update specialist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecialistBody"
      responses:
        "200":
          description: Updated specialist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialist"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: childId
          in: query
          schema:
            type: ["integer", "null"]
        - name: specialistId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: difficulty
          in: query
          schema:
            type: ["string", "null"]
        - name: minAge
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxAge
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create exercise (admin/specialist)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseBody"
      responses:
        "201":
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
    patch:
      operationId: updateExercise
      tags: [exercises]
      summary: Update exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExerciseBody"
      responses:
        "200":
          description: Updated exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, specialist, parent, child]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, specialist, parent, child]
        phone:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
    Specialist:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        totalSessions:
          type: integer
        activeChildren:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - email
        - specialty
        - totalSessions
        - activeChildren
        - createdAt
    CreateSpecialistBody:
      type: object
      properties:
        userId:
          type: integer
        specialty:
          type: string
        bio:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
      required:
        - userId
        - specialty
    UpdateSpecialistBody:
      type: object
      properties:
        specialty:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        childId:
          type: integer
        childName:
          type: string
        specialistId:
          type: integer
        specialistName:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - childId
        - childName
        - specialistId
        - specialistName
        - date
        - time
        - duration
        - status
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        childId:
          type: integer
        specialistId:
          type: integer
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        notes:
          type: ["string", "null"]
      required:
        - childId
        - specialistId
        - date
        - time
        - duration
    UpdateAppointmentBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [pronunciation, vocabulary, fluency, listening, breathing]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        minAge:
          type: integer
        maxAge:
          type: integer
        instructions:
          type: string
        duration:
          type: integer
        emoji:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - difficulty
        - minAge
        - maxAge
        - instructions
        - duration
        - emoji
        - createdAt
    CreateExerciseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [pronunciation, vocabulary, fluency, listening, breathing]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        minAge:
          type: integer
        maxAge:
          type: integer
        instructions:
          type: string
        duration:
          type: integer
        emoji:
          type: string
      required:
        - title
        - description
        - category
        - difficulty
        - minAge
        - maxAge
        - instructions
        - duration
        - emoji
    UpdateExerciseBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
...